creation Des tables dans MongoDb et MySQL

+++ GenericApp/src/main/java/Emsi/MIAGE/Entities/TypeCapteur.java
@@ -20,7 +20,7 @@ import lombok.ToString;
 @ToString
 public class TypeCapteur {
 	@Id
-   private int typeCapteur;
+   private Long typeCapteur;
    private String libelle;
 
 }
